rfid-service/requirements.txt
requests==2.31.0
RPi.GPIO==0.7.1
mfrc522==0.0.7
smbus2==0.4.3
RPLCD==1.3.1
